fix: Hardcode known intermediary thumbprints in github oidc provider
Following GitHub's announcement that can be found on their [blog](https://github.blog/changelog/2023-06-27-github-actions-update-on-oidc-integration-with-aws/)
there are now multiple thumbprints that are required to integrate
AWS with GitHub actions.

We have been running with this change for some time now and wanted
to submit upstream in case others are having issues with their
GitHub actions workflows assuming AWS roles.