Allow user to change own password when no MFA is present
This allows a user to change their own password when no MFA is present on the account. This is only needed in once situation: when a new users signs in, and is then forced to choose a new password. The MFA isn't set up at that point, and the changing of the password is blocked.